Project document 3: responsive design, adding commas
