+++ calculator-app/src/App.css
@@ -5,6 +5,7 @@
   justify-content: center;
   gap: 1rem;
   width: 100%;
+  margin: auto;
 }
 
 .header {
@@ -24,6 +25,7 @@
 
 @media screen and (width < 550px) {
   .calculator {
-    width: 90%;
+    width: 73%;
+    min-width: 300px;
   }
 }
