+++ calculator-app/src/index.css
@@ -67,14 +67,6 @@
   --light-keys-hover: hsl(290, 56%, 39%);
 }
 
-a {
-  color: var(--text-color);
-}
-
-footer {
-  font-size: 12px;
-}
-
 body {
   display: flex;
   flex-direction: column;
@@ -93,6 +85,12 @@ body {
 
 .attribution {
   text-align: center;
+  font-size: 12px;
+  color: var(--text-color-white);
+}
+
+a {
+  color: var(--text-color-white);
 }
 
 @media screen and (width < 550px) {
@@ -102,13 +100,6 @@ body {
     height: 100%;
   }
 }
-
-/* @media screen and (width > 550px) {
-  button {
-    width: 90px;
-    height: 60px;
-  }
-} */
 
 /* @media (prefers-color-scheme: dark) {
   :root {
