+++ calculator-app/src/utils/keypadLayout.js
@@ -19,3 +19,6 @@ export const buttonsLayout = [
   "RESET", // clear inputs
   "=",
 ];
+
+export const numericBtn = ["0", "1", "2", "3", "4", "5", "6", "7", "8", "9"];
+export const arithmaticeBtn = ["+", "-", "×", "/", "."];
